java_agent/requirements.txt
google-genai
ollama
gitpython
streamlit
python-dotenv
